backend/requirements_stable.txt
# Core FastAPI and web framework
fastapi==0.116.1
uvicorn[standard]==0.24.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
starlette==0.47.3

# Google Cloud services (compatible versions)
google-cloud-firestore==2.21.0
google-cloud-bigquery==3.36.0
google-cloud-storage==2.19.0
google-cloud-translate==3.21.1
google-cloud-aiplatform==1.71.1
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-audit-log==0.3.2

# AI and Machine Learning
vertexai==1.71.1
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15

# Firebase Auth
firebase-admin==6.2.0

# Data validation and serialization
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic-core==2.33.2

# HTTP clients
httpx==0.28.1
requests==2.32.5
aiohttp==3.9.5
httpcore==1.0.9
httplib2==0.30.0

# Google API dependencies
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-api-core==2.25.1
google-api-python-client==2.181.0
googleapis-common-protos==1.70.0
google-resumable-media==2.7.2
google-crc32c==1.7.1
grpcio==1.74.0
grpcio-status==1.62.3
grpc-google-iam-v1==0.14.2
proto-plus==1.26.1
protobuf==4.25.8

# Image & media processing
opencv-python==4.11.0.86
Pillow==10.4.0
ImageHash==4.3.2
PyExifTool==0.5.6
pytesseract==0.3.13
ffmpeg-python==0.2.0

# Machine Learning and Data Science
faiss-cpu==1.12.0
numpy==1.26.4
pandas==2.3.2
scikit-learn==1.7.1
scipy==1.16.1
joblib==1.5.2
threadpoolctl==3.6.0

# Web scraping
feedparser==6.0.11
beautifulsoup4==4.13.5
lxml==6.0.1
soupsieve==2.8

# Auth & security
PyJWT==2.10.1
cryptography==45.0.7
bcrypt==4.3.0
ecdsa==0.19.1

# Caching
redis==6.4.0

# Background tasks
celery==5.4.0
kombu==5.5.4
billiard==4.2.1
amqp==5.3.1
vine==5.1.0

# Monitoring and logging
structlog==25.4.0
prometheus-client==0.22.1

# Development and testing
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
coverage==7.10.6

# Code quality tools
black==25.1.0
isort==6.0.1
flake8==7.3.0
mypy==1.17.1
pre-commit==4.3.0
pycodestyle==2.14.0
pyflakes==3.4.0
mccabe==0.7.0

# Utility libraries
email-validator==2.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
click==8.2.1
colorama==0.4.6
humanize==4.13.0
tqdm==4.67.1
packaging==24.2
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
six==1.17.0
typing-extensions==4.15.0

# Async support
anyio==4.10.0
sniffio==1.3.1
h11==0.16.0
websockets==15.0.1
watchfiles==1.1.0

# Additional dependencies
marshmallow==3.26.1
dataclasses-json==0.6.7
orjson==3.11.3
PyYAML==6.0.2
jsonpatch==1.33
jsonpointer==3.0.0
attrs==25.3.0
frozenlist==1.7.0
multidict==6.6.4
yarl==1.20.1
aiosignal==1.4.0
propcache==0.3.2
